requirements.txt
langchain
langchain-google-genai
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib